requirements.txt
git+https://github.com/tercen/tercen_python_client@0.10.3
python-pptx==0.6.*
python-docx==1.1.*
pillow==10.2.*
